# docs/openapi/swagger.yaml
openapi: "3.0.3"
info:
  title: "STAIR Platform API"
  description: "API for stair configuration, pricing, and manufacturing"
  version: "1.0.0"
servers:
  - url: "http://localhost:8080/api/v1"
    description: "Local development"
  - url: "https://api.stairplatform.com/api/v1"
    description: "Production"
...
paths:
...
  /storage/{key}:
    delete:
      tags: ["projects"]
      summary: "Удалить {key}"
      security:
        - bearerAuth: []
        - cookieAuth: []
        - csrfToken: []
      parameters:
        - name: key
          in: path
          required: true
          description: "Идентификатор key"
          schema:
            type: string
      responses:
        204:
          description: "OK"
          content:
            application/json:
              schema:
                type: object
        401:
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: "Forbidden (в т.ч. CSRF-токен)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        429:
          description: "Rate limit exceeded"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        400:
          description: "Invalid request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: "Not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        422:
          description: "Validation error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: "Internal error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      tags: ["projects"]
      summary: "Получить {key}"
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: key
          in: path
          required: true
          description: "Идентификатор key"
          schema:
            type: string
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                type: object
        401:
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        429:
          description: "Rate limit exceeded"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        400:
          description: "Invalid request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: "Not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        422:
          description: "Validation error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: "Internal error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # Нестандартные RPC-маршруты с ":" в пути (см. x-rpc-routes) не могут быть ключами paths в OpenAPI 3.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: "bearer"
      description: "API-ключ (Authorization: Bearer) — EDR-0016"
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: "Сессионная cookie (httpOnly, SameSite=Strict); для admin — session_admin"
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: "Double-submit CSRF: значение cookie csrf (csrf_admin) должно быть продублировано в этом заголовке для всех изменяющих запросов"
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: "Машинный код ошибки (not_found, forbidden, validation_error, invalid_input, rates_not_allowed, csrf, internal и др.)"
            message:
              type: string
              description: Человекочитаемое сообщение
            request_id:
              type: string
              description: "Идентификатор запроса (только для 5xx)"
        request_id:
          type: string
          description: "Идентификатор запроса для поддержки"
...